+++ src/components/Counter/Counter.tsx
@@ -32,32 +32,32 @@ const Counter: React.FC<Props> = ({
   suffix,
   iconBefore
 }) => {
+  const [counterValue, setCounterValue]: any = React.useState(value);
+
   const fixedNumber = countDecimals(steps);
   const updateValue = (val: number) => {
+    setCounterValue(val);
     onChange(val, suffix);
   };
 
   const handleChange = (event: HandleChange) => {
     const { value: eventValue, validity } = event.target;
-    const numberValue: number = parseFloat(eventValue);
 
     if (validity.valid) {
-      if (eventValue !== '') {
-        updateValue(parseFloat(numberValue.toFixed(fixedNumber)));
-      } else {
-        updateValue(0);
-      }
+      setCounterValue(eventValue);
     }
   };
 
   const handleBlur = (event: HandleChange) => {
     const { value: eventValue } = event.target;
-    const numberValue: number = parseFloat(eventValue);
+    const numberValue: number | string = parseFloat(eventValue);
 
     if (numberValue > max) {
       updateValue(max);
-    } else if (numberValue < min) {
+    } else if (numberValue < min || Number.isNaN(numberValue)) {
       updateValue(min);
+    } else {
+      updateValue(numberValue);
     }
   };
 
@@ -70,12 +70,12 @@ const Counter: React.FC<Props> = ({
           type="button"
           onClick={() =>
             updateValue(
-              value - steps >= min
-                ? parseFloat((value - steps).toFixed(fixedNumber))
+              counterValue - steps >= min
+                ? parseFloat((counterValue - steps).toFixed(fixedNumber))
                 : min
             )
           }
-          disabled={value === min}
+          disabled={counterValue === min}
           aria-label="Decrease"
         >
           <svg viewBox="0 0 24 24" width="24" height="24">
@@ -86,7 +86,7 @@ const Counter: React.FC<Props> = ({
           <input
             type="text"
             pattern="[0-9.]*"
-            value={value}
+            value={counterValue}
             onChange={handleChange}
             onBlur={handleBlur}
           />
@@ -96,12 +96,12 @@ const Counter: React.FC<Props> = ({
           type="button"
           onClick={() =>
             updateValue(
-              value + steps <= max
-                ? parseFloat((value + steps).toFixed(fixedNumber))
+              counterValue + steps <= max
+                ? parseFloat((counterValue + steps).toFixed(fixedNumber))
                 : max
             )
           }
-          disabled={value === max}
+          disabled={counterValue === max}
           aria-label="Increase"
         >
           <svg viewBox="0 0 24 24" width="24" height="24">